+++ src/JsonApi/Serializer/SerializerInterface.php
@@ -7,7 +7,7 @@ use Psr\Http\Message\ResponseInterface;
 
 interface SerializerInterface
 {
-    public function serialize(ResponseInterface $response, int $responseCode, array $content): ResponseInterface;
+    public function serialize(ResponseInterface $response, array $content): ResponseInterface;
 
     public function getBodyAsString(ResponseInterface $response): string;
 }