+++ tests/JsonApi/Double/StubErrorDocument.php
@@ -7,7 +7,6 @@ use WoohooLabs\Yin\JsonApi\Schema\Document\AbstractErrorDocument;
 use WoohooLabs\Yin\JsonApi\Schema\Error\Error;
 use WoohooLabs\Yin\JsonApi\Schema\JsonApiObject;
 use WoohooLabs\Yin\JsonApi\Schema\Link\DocumentLinks;
-use WoohooLabs\Yin\JsonApi\Schema\Link\ErrorLinks;
 
 class StubErrorDocument extends AbstractErrorDocument
 {