+++ tests/JsonApi/Serializer/JsonSerializerTest.php
@@ -12,44 +12,31 @@ class JsonSerializerTest extends TestCase
     /**
      * @test
      */
-    public function setResponseCode()
-    {
-        $serializer = new JsonSerializer();
-
-        $response = $serializer->serialize(new Response(), 418, []);
-
-        $this->assertEquals(418, $response->getStatusCode());
-    }
-
-    /**
-     * @test
-     */
-    public function setContentHeader()
-    {
-        $serializer = new JsonSerializer();
-
-        $response = $serializer->serialize(new Response(), 200, []);
-
-        $this->assertEquals(["application/vnd.api+json"], $response->getHeader("Content-Type"));
-    }
-
-    /**
-     * @test
-     */
     public function serializeBody()
     {
-        $content = [
-            "data" => [
-                "type" => "cat",
-                "id" => "tom",
-            ],
-        ];
-
         $serializer = new JsonSerializer();
 
-        $response = $serializer->serialize(new Response(), 200, $content);
+        $response = $serializer->serialize(
+            new Response(),
+            [
+                "data" => [
+                    "type" => "cat",
+                    "id" => "tom",
+                ],
+            ]
+        );
 
-        $this->assertEquals(json_encode($content), $response->getBody()->__toString());
+        $this->assertEquals(
+            json_encode(
+                [
+                    "data" => [
+                        "type" => "cat",
+                        "id" => "tom",
+                    ],
+                ]
+            ),
+            $response->getBody()->__toString()
+        );
     }
 
     /**